requirements.txt
requests
bs4
pandas
python-dotenv
mysql-connector-python
sqlalchemy
pymysql
pyyaml
great_expectations 